Move src/locales/*.ts to src/locales/index.ts and change import.
